Levels order is now float

+++ src/EcoFarmDataModule/EcoFarmAdmin/EcoFarmAdmin/Domain/Level.cs
@@ -7,7 +7,7 @@ namespace EcoFarmAdmin.Domain;
 public class Level : INotifyPropertyChanged
 {
 	public int     Id         { get; set; }
-	public int     Order      { get; set; }
+	public float   Order      { get; set; }
 	public string? Commentary { get; set; }
 
 	public event PropertyChangedEventHandler? PropertyChanged;
